Reorganisation and renaming
Renamed inputwindow.py to event_viewer.py.
Renamed outputwindow.py to event_sender.py.
Moved some more midi related methods from helpers.py to midihandler.py.
Moved some less midi related methods from midihandler.py to helpers.py.
