+++ helpers.py
@@ -10,6 +10,7 @@ Description of file
 """
 
 
+import subprocess
 import unicodedata
 
 
@@ -40,46 +41,6 @@ def sort_by_col(tree, col, descending):
 	)
 
 
-def event_parser(event):
-
-	details = event[0]
-	timestamp = event[1]
-
-	byte0 = details[0]
-	byte1 = details[1]
-	byte2 = details[2]
-
-	message_code = (byte0 & 0xf0) >> 4
-	message = lookup_message(message_code)
-
-	channel = (byte0 & 0x0f) + 1
-
-	note = byte1
-	velocity = byte2
-
-	return [timestamp, message, channel, note, velocity]
-
-
-def lookup_message(message_code):
-
-	if message_code == 0x8:
-		return "Note Off"
-	elif message_code == 0x9:
-		return "Note On"
-	elif message_code == 0xa:
-		return "Poly Key Pressure"
-	elif message_code == 0xb:
-		return "Controller Change"
-	elif message_code == 0xc:
-		return "Program Change"
-	elif message_code == 0xd:
-		return "Channel Pressure"
-	elif message_code == 0xe:
-		return "Pitch Bend"
-	else:
-		return "???"
-
-
 def is_number(value):
 
 	try:
@@ -103,3 +64,13 @@ def change_data_to_num(data):
 		data[ix] = (float(item[0]), item[1])
 
 	return data
+
+
+def open_event_viewer(id, name):
+
+	subprocess.Popen(["python", "event_viewer.py", str(id), name])
+
+
+def open_event_sender(id, name):
+
+	subprocess.Popen(["python", "event_sender.py", str(id), name])
